+++ Carthage/Checkouts/JTAppleCalendar/Example/JTAppleCalendar iOS Example/ExampleDateCells/DateCellCreatedWithXIB/CellView.swift
@@ -10,7 +10,7 @@
 import JTAppleCalendar
 
 class CellView: JTAppleCell {
-    @IBOutlet var selectedView: AnimationView!
+    @IBOutlet var selectedView: UIView!
     @IBOutlet var dayLabel: UILabel!
     @IBOutlet var monthLabel: UILabel!
 }